int fix for preact_resnet

+++ models/preact_resnet.py
@@ -46,11 +46,11 @@ class PreActWideResNet(nn.Module):
 
         self.conv1 = nn.Conv2d(3, self.in_planes, kernel_size=3, stride=1, padding=1, bias=False)
         self.bn = nn.BatchNorm2d(self.in_planes)
-        self.layer1 = self._make_layer(block, self.first_block*width, num_blocks[0], stride=1)
-        self.layer2 = self._make_layer(block, self.first_block*width*2, num_blocks[1], stride=2)
-        self.layer3 = self._make_layer(block, self.first_block*width*4, num_blocks[2], stride=2)
+        self.layer1 = self._make_layer(block, int(self.first_block*width), num_blocks[0], stride=1)
+        self.layer2 = self._make_layer(block, int(self.first_block*width*2), num_blocks[1], stride=2)
+        self.layer3 = self._make_layer(block, int(self.first_block*width*4), num_blocks[2], stride=2)
 
-        self.linear = nn.Linear(self.first_block*width*4, num_classes)
+        self.linear = nn.Linear(int(self.first_block*width*4), num_classes)
 
     def _make_layer(self, block, planes, num_blocks, stride):
         strides = [stride] + [1]*(num_blocks-1)
